+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Trunk
 
+* examples: new mixed output stream example
 * handler: bind execution context with current instance
 
 ## Version 1.0.4